refactor: extract expect_revert cheatcode's context to separate class
